appointments-service/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
requests==2.31.0
python-dateutil==2.8.2 